+++ packages/oxlint-tailwindcss/src/rules/enforce-consistent-line-wrapping.ts
@@ -443,6 +443,6 @@ export const enforceConsistentLineWrapping = defineRule({
       }
     }
 
-    return createExtractorVisitors(context, check)
+    return createExtractorVisitors(context, check, { raw: true })
   },
 })